# Bulk get cached projection