--- a/docker-compose.yml
+++ b/docker-compose.yml
@@ -28,7 +28,8 @@ services:
     ports: ["6379:6379"]
     volumes:
       - "redis_data:/data"
-    command: "valkey-server --save 30 1 --loglevel warning"
+      - "./docker/valkey.conf:/etc/valkey.conf"
+    command: "valkey-server /etc/valkey.conf --save 30 1 --loglevel warning"
     healthcheck:
       test: ["CMD", "valkey-cli", "ping"]
       start_period: 4s