--- a/docker/valkey.conf
+++ b/docker/valkey.conf
@@ -1,0 +1,2 @@
+maxmemory-policy allkeys-lru
+maxmemory 1gb