## Description

The idea is to bulk-get the cached projection to limit the number of queries performed to Redis.

## Benchmark

Condition:
- I performed the test when everything was cached.
- It was tested locally
- Timetable size: 823 trains
- Valid trains: 674

The `path_projection` performance doesn't seem to change (1.2s).

Queries performances:
- Before: 1 query took ~0,07ms (times 674 = 47ms)
- Now: 1 bulk query takes ~8ms

I think in production, because of queries overhead the difference will be far more important.


The number of Redis queries is reduced by the number of valid train schedule. (2211 -> 1532 queries).
Next, I will do the same to retrieve the path and simulation from the cache (but this implies code refactoring).

> [!IMPORTANT]
> There is a limitation with this. I didn't find how to set the expiration time of Redis entry with a bulk command. So now the Redis entry expires after 1 week even if they're used recently.