@flomonster 2024-07-08
> There is a limitation with this. I didn't find how to set the expiration time of Redis entry with a bulk command. So now the Redis entry expires after 1 week even if they're used recently.

I found a solution. We can use a transaction so all the commands are executed at once. Example with the CLI:

```
$ MSET hey 42 hi 43
$ MULTI
$ EXPIRE hey 10
$ EXPIRE hi 10
$ EXEC
```