@flomonster 2024-07-08
A better solution is to set up an `allkeys-lru` policy to let Redis handle this automatically.